Move xover suggestion algorithm to a separate file
