+++ ensnano-design/src/lib.rs
@@ -423,6 +423,15 @@ impl Design {
     pub fn prepare_for_save(&mut self, saving_information: SavingInformation) {
         self.saved_camera = saving_information.camera;
     }
+
+    pub fn get_nucl_position(&self, nucl: Nucl) -> Option<Vec3> {
+        let helix = self.helices.get(&nucl.helix)?;
+        Some(helix.space_pos(
+            &self.parameters.unwrap_or_default(),
+            nucl.position,
+            nucl.forward,
+        ))
+    }
 }
 
 pub struct SavingInformation {
