+++ src/app_state/design_interactor/presenter/design_content.rs
@@ -30,6 +30,9 @@ use ultraviolet::Vec3;
 
 use grid_data::GridManager;
 
+mod xover_suggestions;
+use xover_suggestions::XoverSuggestions;
+
 #[derive(Default, Clone)]
 pub(super) struct DesignContent {
     /// Maps identifer of elements to their object type
@@ -51,10 +54,6 @@ pub(super) struct DesignContent {
     /// Maps the identifier of an element to its color
     pub color: HashMap<u32, u32, RandomState>,
     pub basis_map: Arc<HashMap<Nucl, char, RandomState>>,
-    /// The position in space of the nucleotides in the Red group
-    pub red_cubes: HashMap<(isize, isize, isize), Vec<Nucl>, RandomState>,
-    /// The list of nucleotides in the blue group
-    pub blue_nucl: Vec<Nucl>,
     pub prime3_set: Vec<Prime3End>,
     pub elements: Vec<DnaElement>,
     pub suggestions: Vec<(Nucl, Nucl)>,
@@ -284,94 +283,6 @@ impl DesignContent {
             .map(|t| *t.0)
             .collect()
     }
-
-    /// Return the list of all suggested crossovers
-    fn get_suggestions(
-        &self,
-        design: &Design,
-        suggestion_parameters: &SuggestionParameters,
-    ) -> Vec<(Nucl, Nucl)> {
-        let mut ret = vec![];
-        for blue_nucl in self.blue_nucl.iter() {
-            let neighbour = self
-                .get_possible_cross_over(design, blue_nucl, suggestion_parameters)
-                .unwrap_or_default();
-            for (red_nucl, dist) in neighbour {
-                ret.push((*blue_nucl, red_nucl, dist))
-            }
-        }
-        ret.sort_by(|a, b| a.2.partial_cmp(&b.2).unwrap());
-        self.trimm_suggestion(&ret)
-    }
-
-    /// Trimm a list of crossovers so that each nucleotide appears at most once in the suggestion
-    /// list.
-    fn trimm_suggestion(&self, suggestion: &[(Nucl, Nucl, f32)]) -> Vec<(Nucl, Nucl)> {
-        let mut used = HashSet::new();
-        let mut ret = vec![];
-        for (a, b, _) in suggestion {
-            if !used.contains(a) && !used.contains(b) {
-                ret.push((*a, *b));
-                used.insert(a);
-                used.insert(b);
-            }
-        }
-        ret
-    }
-
-    /// Return all the crossovers of length less than `len_crit` involving `nucl`, and their length.
-    fn get_possible_cross_over(
-        &self,
-        design: &Design,
-        nucl: &Nucl,
-        suggestion_parameters: &SuggestionParameters,
-    ) -> Option<Vec<(Nucl, f32)>> {
-        let mut ret = Vec::new();
-        let positions = self
-            .identifier_nucl
-            .get(nucl)
-            .and_then(|id| self.space_position.get(id))?;
-        let cube0 = space_to_cube(positions[0], positions[1], positions[2]);
-
-        let len_crit = 1.2;
-        for i in vec![-1, 0, 1].iter() {
-            for j in vec![-1, 0, 1].iter() {
-                for k in vec![-1, 0, 1].iter() {
-                    let cube = (cube0.0 + i, cube0.1 + j, cube0.2 + k);
-                    if let Some(v) = self.red_cubes.get(&cube) {
-                        for red_nucl in v {
-                            if red_nucl.helix != nucl.helix {
-                                if let Some(red_position) = self
-                                    .identifier_nucl
-                                    .get(&red_nucl)
-                                    .and_then(|id| self.space_position.get(id))
-                                {
-                                    let dist = (0..3)
-                                        .map(|i| (positions[i], red_position[i]))
-                                        .map(|(x, y)| (x - y) * (x - y))
-                                        .sum::<f32>()
-                                        .sqrt();
-                                    if dist < len_crit
-                                        && (suggestion_parameters.include_scaffold
-                                            || design.get_strand_nucl(nucl) != design.scaffold_id)
-                                        && (suggestion_parameters.include_scaffold
-                                            || design.get_strand_nucl(red_nucl)
-                                                != design.scaffold_id)
-                                        && (suggestion_parameters.include_intra_strand
-                                            || design.get_strand_nucl(nucl)
-                                                != design.get_strand_nucl(red_nucl))
-                                    {
-                                        ret.push((*red_nucl, dist));
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        Some(ret)
-    }
 }
 
 #[derive(Debug)]
@@ -417,11 +328,10 @@ impl DesignContent {
         let mut nucl_id;
         let mut old_nucl: Option<Nucl> = None;
         let mut old_nucl_id = None;
-        let mut red_cubes = HashMap::default();
         let mut elements = Vec::new();
         let mut prime3_set = Vec::new();
-        let mut blue_nucl = Vec::new();
         let mut new_junctions: JunctionsIds = Default::default();
+        let mut suggestion_maker = XoverSuggestions::default();
         xover_ids.agree_on_next_id(&mut new_junctions);
         let mut grid_manager = GridManager::new_from_design(&design);
         if *old_grid_ptr != Some(Arc::as_ptr(&design.grids) as usize) {
@@ -495,16 +405,7 @@ impl DesignContent {
                             basis_map.remove(&nucl);
                         }
                         strand_position += 1;
-                        match groups.get(&nucl.helix) {
-                            Some(true) => {
-                                blue_nucl.push(nucl);
-                            }
-                            Some(false) => {
-                                let cube = space_to_cube(position.x, position.y, position.z);
-                                red_cubes.entry(cube).or_insert(vec![]).push(nucl.clone());
-                            }
-                            None => (),
-                        }
+                        suggestion_maker.add_nucl(nucl, position, groups.as_ref());
                         let position = [position[0] as f32, position[1] as f32, position[2] as f32];
                         space_position.insert(nucl_id, position);
                         if let Some(old_nucl) = old_nucl.take() {
@@ -602,14 +503,12 @@ impl DesignContent {
             color: color_map,
             helix_map,
             basis_map: Arc::new(basis_map),
-            red_cubes,
             prime3_set,
-            blue_nucl,
             elements,
             grid_manager,
             suggestions: vec![],
         };
-        let suggestions = ret.get_suggestions(&design, suggestion_parameters);
+        let suggestions = suggestion_maker.get_suggestions(&design, suggestion_parameters);
         ret.suggestions = suggestions;
 
         drop(groups);
@@ -658,15 +557,6 @@ impl DesignContent {
     pub fn read_simualtion_update(&mut self, update: &dyn SimulationUpdate) {
         update.update_positions(&self.identifier_nucl, &mut self.space_position)
     }
-}
-
-fn space_to_cube(x: f32, y: f32, z: f32) -> (isize, isize, isize) {
-    let cube_len = 1.2;
-    (
-        x.div_euclid(cube_len) as isize,
-        y.div_euclid(cube_len) as isize,
-        z.div_euclid(cube_len) as isize,
-    )
 }
 
 #[cfg(test)]
